+++ lib/lightning/config/bootstrap.ex
@@ -388,12 +388,6 @@ defmodule Lightning.Config.Bootstrap do
           port: port,
           compress: true,
           protocol_options: [
-            max_frame_size:
-              Application.get_env(
-                :lightning,
-                :max_dataclip_size_bytes,
-                10_000_000
-              ),
             # Note that if a request is more than 10x the max dataclip size, we cut
             # the connection immediately to prevent memory issues via the
             # :max_skip_body_length setting.